fix(knowledge-base): provider-route ask synthesis (#12741)
