+++ ai/services/memory-core/MemoryService.mjs
@@ -3,7 +3,8 @@ import StorageRouter         from './managers/StorageRouter.mjs';
 import crypto                from 'crypto';
 import GraphService          from './GraphService.mjs';
 import logger                from '../../mcp/server/memory-core/logger.mjs';
-import SessionService, {buildChatModel} from './SessionService.mjs';
+import SessionService        from './SessionService.mjs';
+import {buildChatModel}      from '../../provider/buildChatModel.mjs';
 import aiConfig              from '../../mcp/server/memory-core/config.mjs';
 import RequestContextService, {SHARED_USER_ID, normalizeUserId} from '../../mcp/server/shared/services/RequestContextService.mjs';
 import {IDENTITIES, TRUST_TIERS, TRUST_TIER_ORDER} from '../../graph/identityRoots.mjs';
